Suppression du type BigDecimal

+++ src/main/java/nc/noumea/mairie/model/repository/IMairieRepository.java
@@ -1,6 +1,5 @@
 package nc.noumea.mairie.model.repository;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import nc.noumea.mairie.model.bean.Spadmn;
@@ -9,17 +8,17 @@ import nc.noumea.mairie.model.bean.Spmtsr;
 
 public interface IMairieRepository {
 
-	List<BigDecimal> getListeCarriereActiveAvecPAAffecte();
+	List<Integer> getListeCarriereActiveAvecPAAffecte();
 
-	List<BigDecimal> getListeCarriereActiveAvecPA();
-	
+	List<Integer> getListeCarriereActiveAvecPA();
+
 	Spadmn chercherPositionAdmAgentAncienne(Integer noMatr);
-	
+
 	Spadmn chercherPositionAdmAgentEnCours(Integer noMatr);
-	
+
 	List<Spmtsr> getListSpmtsr(Integer noMatricule);
-	
+
 	Spcarr getCarriereFonctionnaireAncienne(Integer noMatr);
-	
+
 	Spcarr getCarriereActive(Integer noMatr);
 }
